+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file documents the changes made to the formatter with each release.
 
+## Unreleased
+
+### Added
+
+- Linter: add support for stdin (#327)
+
+### Fixed
+
+- Linter: respect the EditorConfig indentation size when checking line lengths (#318)
+
 ## Release 0.27.0 (2026-09-25)
 
 ### Added